requirements.txt
faiss-cpu
ollama
numpy
pypdf
